refactor : 모달 렌더링 최적화 구조 변경
