+++ src/1_Page/Team/ui/MakeMatchBtn/index.tsx
@@ -1,0 +1,33 @@
+import MakeTeamMatchModal from "./ui/MakeTeamMatchModal";
+import React from "react";
+
+const MakeMatchBtn = (props: string) => {
+  const { team_list_idx } = props;
+  const [isMakeTeamMatchModal, setIsMakeTeamMatchModal] = React.useState(false);
+  const toggleMakeTeamMatchModal = () =>
+    setIsMakeTeamMatchModal(!isMakeTeamMatchModal);
+
+  return (
+    <div className="flex gap-2">
+      <button className="bg-blue-500 text-white text-sm font-medium py-1 px-3 rounded-full">
+        팀관리
+      </button>
+      <button
+        className="bg-blue-500 text-white text-sm font-medium py-1 px-3 rounded-full cursor-auto"
+        onClick={() => {
+          setIsMakeTeamMatchModal(true);
+        }}>
+        매치 생성
+      </button>
+      {/* 매치 생성 모달 */}
+      {isMakeTeamMatchModal && (
+        <MakeTeamMatchModal
+          team_list_idx={team_list_idx}
+          onClose={toggleMakeTeamMatchModal}
+        />
+      )}
+    </div>
+  );
+};
+
+export default MakeMatchBtn;
