+++ src/1_Page/Team/ui/MakeMatchBtn/ui/MakeTeamMatchModal/index.tsx
@@ -1,18 +1,19 @@
-import { TeamListIdxProps } from "./type";
+import { makeTeamMatchModalProps } from "./type";
 
-import usePostTeamMatch from "../../../../3_Entity/Match/usePostTeamMatch";
-import { teamMatchAttribute } from "../../../../4_Shared/constant/teamMatchAttribute";
-import { matchType } from "../../../../4_Shared/constant/matchType";
-import { matchParticipation } from "../../../../4_Shared/constant/matchParticipation";
-import { matchDuration } from "../../../../4_Shared/constant/matchDuration";
-import { formation } from "../../../../4_Shared/constant/formation";
+import usePostTeamMatch from "../../../../../../3_Entity/Match/usePostTeamMatch";
+import { teamMatchAttribute } from "../../../../../../4_Shared/constant/teamMatchAttribute";
+import { matchType } from "../../../../../../4_Shared/constant/matchType";
+import { matchParticipation } from "../../../../../../4_Shared/constant/matchParticipation";
+import { matchDuration } from "../../../../../../4_Shared/constant/matchDuration";
+import { formation } from "../../../../../../4_Shared/constant/formation";
 import { ExtendedMatchFormData } from "./type";
 import { useForm, SubmitHandler } from "react-hook-form";
 
-import { MatchFormData } from "../../../../3_Entity/Match/type";
+import { MatchFormData } from "../../../../../../3_Entity/Match/type";
 import { validateTime } from "./util/validate";
 
-const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
+const MakeTeamMatchModal = (props: makeTeamMatchModalProps) => {
+  const { team_list_idx, onClose } = props;
   const today = new Date();
   const formattedDate = today.toISOString().split("T")[0];
   const formattedTime = today.toTimeString().slice(0, 5);
@@ -37,10 +38,7 @@ const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
     } as ExtendedMatchFormData,
   });
 
-  const selectedMatchType = watch("match_type_idx_radio");
-  console.log(selectedMatchType);
-
-  const isCanFormationChange = selectedMatchType === "0";
+  const isCanFormationChange = watch("match_type_idx_radio") === "0";
 
   const onSubmit: SubmitHandler<ExtendedMatchFormData> = (data) => {
     const {
@@ -49,7 +47,7 @@ const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
       match_type_idx_radio,
       match_match_participation_type_radio,
       ...rest
-    } = data;
+    } = data; // 입력 폼과 실제 데이터가 다르기 때문에 분리
     const validationError = validateTime(
       match_match_start_date,
       match_match_start_time
@@ -72,7 +70,6 @@ const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
       match_type_idx: Number(match_type_idx_radio),
       match_match_attribute: Number(data.match_match_attribute),
       match_match_start_time: `${match_match_start_date} ${match_match_start_time}`,
-      match_match_duration: data.match_match_duration,
     };
     postEvent(submitData);
   };
@@ -82,7 +79,11 @@ const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
       <div className="bg-white rounded-2xl shadow-xl w-[500px] p-6">
         <div className="flex justify-between items-center mb-6">
           <h2 className="text-xl font-semibold">팀 매치 모달 생성</h2>
-          <button className="text-gray-400 hover:text-gray-600">✖</button>
+          <button
+            className="text-gray-400 hover:text-gray-600"
+            onClick={onClose}>
+            ✖
+          </button>
         </div>
 
         <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
@@ -210,6 +211,7 @@ const MakeTeamMatchModal = ({ team_list_idx }: TeamListIdxProps) => {
           {/* 버튼 */}
           <div className="flex justify-end gap-4 mt-6">
             <button
+              onClick={onClose}
               type="button"
               className="px-4 py-2 rounded-lg border border-gray-300">
               취소
