+++ src/1_Page/Team/ui/MakeMatchBtn/ui/MakeTeamMatchModal/util/validate.ts
