--- src/1_Page/Team/ui/MakeTeamMatchModal/util/convertNum.ts
@@ -1,9 +1,0 @@
-export const convertToNumber = (data: any, fields: string[]) => {
-  const result = { ...data };
-
-  fields.forEach((field) => {
-    result[field] = Number(data[field]);
-  });
-
-  return result;
-};
